Added basic Caching
Added a template metafunction for string reports (cache, borb.py)
moved to service factory view for speech services (speech.py)
sounds can now play streams and files

+++ src/sound.py
@@ -5,7 +5,12 @@ from conf import conf
 
 TMP_BUF = p(conf.cache) / 'tmp.mp3'
 
-def play(byte_stream):
-    TMP_BUF.write_bytes(byte_stream)
-    playsound(TMP_BUF.__str__())
-    TMP_BUF.unlink()
+def play(resource,):
+    if isinstance(resource, p):
+        playsound(resource.__str__())
+    elif isinstance(resource, bytes):
+        TMP_BUF.write_bytes(resource)
+        playsound(TMP_BUF.__str__())
+        TMP_BUF.unlink()
+    else:
+        raise NotImplementedError
